cleanup file structure and router
