+++ src/applications/_mock-form-ae-design-patterns/vadx/server/constants/mockServerPaths.js
@@ -1,0 +1,44 @@
+const MOCK_SERVER_PATHS = [
+  'src/applications/_mock-form-ae-design-patterns/mocks/server.js',
+  'src/applications/appeals/shared/tests/mock-api.js',
+  'src/applications/avs/api/mocks/index.js',
+  'src/applications/check-in/api/local-mock-api/index.js',
+  'src/applications/combined-debt-portal/combined/utils/mocks/mockServer.js',
+  'src/applications/disability-benefits/2346/mocks/index.js',
+  'src/applications/disability-benefits/all-claims/local-dev-mock-api/index.js',
+  'src/applications/education-letters/testing/response.js',
+  'src/applications/financial-status-report/mocks/responses.js',
+  'src/applications/health-care-supply-reordering/mocks/index.js',
+  'src/applications/ivc-champva/10-10D/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/ivc-champva/10-7959C/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/ivc-champva/10-7959f-1/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/mhv-landing-page/mocks/api/index.js',
+  'src/applications/mhv-medications/mocks/api/index.js',
+  'src/applications/mhv-secure-messaging/api/mocks/index.js',
+  'src/applications/mhv-supply-reordering/mocks/index.js',
+  'src/applications/my-education-benefits/testing/responses.js',
+  'src/applications/personalization/dashboard/mocks/server.js',
+  'src/applications/personalization/notification-center/mocks/server.js',
+  'src/applications/personalization/profile/mocks/server.js',
+  'src/applications/personalization/review-information/tests/fixtures/mocks/local-mock-responses.js',
+  'src/applications/post-911-gib-status/mocks/server.js',
+  'src/applications/representative-appoint/mocks/server.js',
+  'src/applications/simple-forms/20-10206/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/simple-forms/20-10207/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/simple-forms/21-0845/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/simple-forms/21-0966/tests/e2e/fixtures/mocks/local-mock-api-responses.js',
+  'src/applications/simple-forms/21-0972/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/simple-forms/21-10210/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/simple-forms/21-4138/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/simple-forms/21-4142/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/simple-forms/40-0247/tests/e2e/fixtures/mocks/local-mock-api-reponses.js',
+  'src/applications/simple-forms/form-upload/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/simple-forms/mock-simple-forms-patterns/tests/e2e/fixtures/mocks/local-mock-responses.js',
+  'src/applications/travel-pay/services/mocks/index.js',
+  'src/applications/vaos/services/mocks/index.js',
+  'src/platform/mhv/api/mocks/index.js',
+  'src/platform/mhv/downtime/mocks/api/index.js',
+  'src/platform/testing/local-dev-mock-api/common.js',
+];
+
+module.exports = { MOCK_SERVER_PATHS };
