+++ src/applications/_mock-form-ae-design-patterns/vadx/server/index.js
@@ -1,10 +1,9 @@
 /* eslint-disable no-console */
 const express = require('express');
 
-const cors = require('./cors');
-const { initializeManifests } = require('./manifests');
-const { autoStartServers } = require('./processes');
-const routes = require('./routes');
+const cors = require('./utils/cors');
+const { initializeManifests } = require('./utils/manifests');
+const { autoStartServers } = require('./utils/processes');
 const parseArgs = require('./utils/parseArgs');
 
 const app = express();
@@ -12,10 +11,21 @@ const port = 1337;
 const args = parseArgs();
 
 app.use(express.json());
-// Allow CORS with pretty open settings
+
+// Allow CORS
 app.use(cors);
 
-app.use(routes);
+const router = express.Router();
+
+router.use(require('./routes/events'));
+router.use(require('./routes/manifests'));
+router.use(require('./routes/output'));
+router.use(require('./routes/start-mock-server'));
+router.use(require('./routes/start-frontend-server'));
+router.use(require('./routes/status'));
+router.use(require('./routes/stop-on-port'));
+
+app.use(router);
 
 app.listen(port, async () => {
   await initializeManifests();
