+++ src/applications/_mock-form-ae-design-patterns/vadx/server/routes/output.js
@@ -1,0 +1,17 @@
+const express = require('express');
+const { outputCache } = require('../utils/processes');
+
+const router = express.Router();
+
+router.get('/output/:name', (req, res) => {
+  const { name } = req.params;
+  if (outputCache[name]) {
+    res.json(outputCache[name]);
+  } else {
+    res
+      .status(404)
+      .json({ error: `No output cache found for process ${name}` });
+  }
+});
+
+module.exports = router;
