+++ src/applications/_mock-form-ae-design-patterns/vadx/server/utils/cors.js
